fix: Mocha test throws SyntaxError
Fixed issue with mocha test throwing syntax errors.
The error is caused by typescript transpiling.
Note that in compiler options, target affects language features in the output code, while module affects the use of `import`/`export`/`require` in the output code. So we had to overwrite module option in the test case so that the transpiled code uses `require` and can be interpreted by mocha.
Build output module option should be unaffected.